client: move common async reed solomon infrastructure to decode-rs
