+++ client/Cargo.toml
@@ -13,7 +13,6 @@ futures = "0.3.25"
 hex = "0.4.3"
 itertools = "0.10.5"
 log = "0.4.17"
-reed-solomon-erasure = { version = "6.0.0", features = ["simd-accel"] }
 reqwest = { version = "0.11.12", features = ["stream"] }
 simple_logger = "4.0.0"
 tokio = { version = "1.21.2", features = ["rt-multi-thread", "macros", "fs"] }
