+++ decode-rs/Cargo.toml
@@ -6,9 +6,11 @@ edition = "2021"
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
+futures = { version = "0.3.25", default-features = false }
 futures-util = "0.3.25"
 hex = "0.4.3"
 log = { version = "0.4.17" }
+reed-solomon-erasure = { version = "6.0.0", features = ["simd-accel"] }
 reqwest = { version = "0.11.12", features = ["json", "stream"] }
 serde = { version = "1.0", features = [ "derive" ] }
 serde_json = { version = "1.0" }
