+++ decode-rs/src/lib.rs
@@ -1,2 +1,4 @@
 pub mod api;
 pub mod api_client;
+pub mod async_reed_solomon;
+pub mod transposed_slices;
